+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Client for the service, to be used by external projects
-
+- Solr Suggest API method
 
 ## [1.0.0] - 2022-05-16
 ### Added